+++ src/frontend/src/sol/schedulers/sol-wallet.scheduler.ts
@@ -12,7 +12,7 @@ import {
 	loadSolLamportsBalance,
 	loadSplTokenBalance
 } from '$sol/api/solana.api';
-import { getSolTransactions } from '$sol/services/sol-transactions.services';
+import { getSolTransactions } from '$sol/services/sol-signatures.services';
 import type { SolCertifiedTransaction } from '$sol/stores/sol-transactions.store';
 import type { SolanaNetworkType } from '$sol/types/network';
 import type { SolBalance } from '$sol/types/sol-balance';
