+++ src/frontend/src/tests/sol/services/sol-transactions.services.spec.ts
@@ -1,9 +1,7 @@
 import { SOLANA_TOKEN_ID } from '$env/tokens/tokens.sol.env';
 import * as solanaApi from '$sol/api/solana.api';
-import {
-	fetchSolTransactionsForSignature,
-	loadNextSolTransactions
-} from '$sol/services/sol-transactions.services';
+import * as solTransactionsServices from '$sol/services/sol-transactions.services';
+import { fetchSolTransactionsForSignature } from '$sol/services/sol-transactions.services';
 import { solTransactionsStore } from '$sol/stores/sol-transactions.store';
 import { SolanaNetworks, type SolanaNetworkType } from '$sol/types/network';
 import type {
@@ -24,14 +22,6 @@ import { mockSolAddress, mockSolAddress2, mockSplAddress } from '$tests/mocks/so
 import { get } from 'svelte/store';
 import { vi, type MockInstance } from 'vitest';
 
-vi.mock(import('$sol/services/sol-transactions.services'), async (importOriginal) => {
-	const actual = await importOriginal();
-	return {
-		...actual,
-		getSolTransactions: vi.fn()
-	};
-});
-
 describe('sol-transactions.services', () => {
 	beforeEach(() => {
 		vi.clearAllMocks();
@@ -41,7 +31,7 @@ describe('sol-transactions.services', () => {
 		solTransactionsStore.reset(SOLANA_TOKEN_ID);
 	});
 
-	describe('fetchSolTransactionsForSignature', () => {
+	describe('fetchSolTransactionsForSignature', async () => {
 		const network: SolanaNetworkType = 'mainnet';
 		const mockSignature: SolSignature = mockSolSignatureResponse();
 		const mockParams = {
@@ -233,13 +223,15 @@ describe('sol-transactions.services', () => {
 					type: 'receive'
 				});
 
-			// spyGetTransactions = vi
-			// 	.spyOn(solTransactionsServices, 'getSolTransactions')
-			// 	.mockResolvedValue(mockTransactions);
+			spyGetTransactions = vi
+				.spyOn(solTransactionsServices, 'getSolTransactions')
+				.mockImplementation(vi.fn());
+
+			// ({ loadNextSolTransactions } = solTransactionsServices);
 		});
 
 		it('should load and return transactions successfully', async () => {
-			// spyGetTransactions.mockResolvedValue(mockTransactions);
+			spyGetTransactions.mockResolvedValue(mockTransactions);
 
 			const transactions = await loadNextSolTransactions({
 				address: mockSolAddress,
@@ -331,122 +323,6 @@ describe('sol-transactions.services', () => {
 				address: mockSolAddress,
 				network: SolanaNetworks.devnet
 			});
-		});
-	});
-
-	describe('getSolTransactions', async () => {
-		const { getSolTransactions } = await vi.importActual<
-			typeof import('$sol/services/sol-transactions.services')
-		>('$sol/services/sol-transactions.services');
-
-		let spyFetchSignatures: MockInstance;
-		let spyFetchTransactionDetailForSignature: MockInstance;
-
-		const mockError = new Error('RPC Error');
-
-		beforeEach(() => {
-			spyFetchSignatures = vi.spyOn(solanaApi, 'fetchSignatures');
-			spyFetchSignatures.mockReturnValue([mockSolSignatureResponse(), mockSolSignatureResponse()]);
-
-			spyFetchTransactionDetailForSignature = vi.spyOn(
-				solanaApi,
-				'fetchTransactionDetailForSignature'
-			);
-			spyFetchTransactionDetailForSignature.mockResolvedValue(mockSolRpcSendTransaction);
-		});
-
-		it('should fetch transactions successfully', async () => {
-			const transactions = await getSolTransactions({
-				address: mockSolAddress,
-				network: SolanaNetworks.mainnet
-			});
-
-			expect(transactions).toHaveLength(2);
-			expect(spyFetchSignatures).toHaveBeenCalledTimes(1);
-			expect(spyFetchTransactionDetailForSignature).toHaveBeenCalledTimes(2);
-		});
-
-		it('should handle before parameter', async () => {
-			const signature = mockSolSignature();
-			await getSolTransactions({
-				address: mockSolAddress,
-				network: SolanaNetworks.mainnet,
-				before: signature
-			});
-
-			expect(spyFetchSignatures).toHaveBeenCalledWith(
-				expect.objectContaining({
-					before: signature
-				})
-			);
-		});
-
-		it('should handle limit parameter', async () => {
-			await getSolTransactions({
-				address: mockSolAddress,
-				network: SolanaNetworks.mainnet,
-				limit: 5
-			});
-
-			expect(spyFetchSignatures).toHaveBeenCalledWith(
-				expect.objectContaining({
-					limit: 5
-				})
-			);
-		});
-
-		it('should not return transactions that do not change SOL balance', async () => {
-			spyFetchTransactionDetailForSignature.mockReturnValue({
-				...mockSolRpcSendTransaction,
-				meta: {
-					...mockSolRpcSendTransaction.meta,
-					postBalances: mockSolRpcSendTransaction.meta?.postBalances,
-					preBalances: mockSolRpcSendTransaction.meta?.postBalances
-				}
-			});
-
-			const transactions = await getSolTransactions({
-				address: mockSolAddress,
-				network: SolanaNetworks.mainnet,
-				limit: 5
-			});
-
-			expect(transactions).toHaveLength(0);
-		});
-
-		it('should handle empty signatures response', async () => {
-			spyFetchSignatures.mockReturnValue([]);
-
-			const transactions = await getSolTransactions({
-				address: mockSolAddress,
-				network: SolanaNetworks.mainnet
-			});
-
-			expect(transactions).toHaveLength(0);
-			expect(spyFetchTransactionDetailForSignature).not.toHaveBeenCalled();
-		});
-
-		it('should handle null transaction responses', async () => {
-			spyFetchSignatures.mockReturnValue([mockSolSignatureResponse()]);
-			spyFetchTransactionDetailForSignature.mockReturnValue(null);
-
-			const transactions = await getSolTransactions({
-				address: mockSolAddress,
-				network: SolanaNetworks.mainnet
-			});
-
-			expect(transactions).toHaveLength(0);
-		});
-
-		it('should handle RPC errors gracefully', async () => {
-			spyFetchSignatures.mockRejectedValue(mockError);
-
-			await expect(
-				getSolTransactions({
-					address: mockSolAddress,
-					network: SolanaNetworks.mainnet
-				})
-			).rejects.toThrow(mockError);
 		});
 	});
 });
